requirements.txt
boto3==1.34.84
holidays==0.18
joblib==1.4.0
matplotlib==3.8.4
numpy==1.26.4
optuna==3.6.1
pandarallel==1.6.5
pandas==2.2.2
plotly==5.19.0
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.32.3
s3fs==2024.6.0
scikit_learn==1.4.2
scipy==1.13.1
shap==0.45.1
swifter
tabulate==0.9.0
tqdm==4.66.1
xgboost==2.0.3
polars==1.4.1
